dev: Make development easier using Nix
If you have Nix (the package manager, not the OS) installed,
with flakes feature enabled, you can enter a dev environment by
running `nix develop`. This will drop you into a shell with all the
dependencies required to build polybar with all the features enabled.

These dependencies are also locked in flake.lock, so this build
environment will be 100% reproducible regardless of the machine you are
running.

It also makes life easier for new devs because now they don't have to
figure out what the dependencies are, which versions of them will work
and what their package names are in ther distro.