feat(audio): Allow label-volume in muted format
Until c552df3b6603 ("fix: Modules did not validate tags used in
formats") it was possible to use the volume label in the muted format.
This made sense as the volume level can change while muted and it
allowed for example following formats:

    format-volume = "<ramp-volume>  <label-volume> "
    format-muted  =  <label-muted> (<label-volume>)

Add label-volume to the format-muted allowlist to re-enable this
feature.