fix(file_util::pick): ensure files are readable before picking them
the firmware of the Lenovo X13s does not expose charge_now or
charge_full, but the files still exist, so file_util::pick() will still
choose that over energy_now/energy_full. This leads to the battery
module erroneously outputting "0%" at all times. If charge_now or
charge_full is read, it will return ENODATA, so simply reading 1 char
and checking the status should determine if the file should actually be
picked.