fix: Stop counting iowait towards CPU utilization
If the cpu is spending a lot of time in iowait, it is still idle but
some process is waiting on io. The system itself is not under load.

This is especially annoying with ghostty, as it spends most time waiting
on io_uring, which drives up the CPU usage reported by polybar to around
50% or more. Other system monitors, such as htop or btop, do not count
iowait towards utilization and only report ~5% CPU utilization instead.

This is discussed in more detail at
https://github.com/ghostty-org/ghostty/discussions/3224.

The kde system monitor was recently adjusted to behave similarly:
https://invent.kde.org/plasma/ksystemstats/-/merge_requests/104/.