battery: fixes to consumption/remaining math
Measurements of current on Linux may be negative to indicate a
discharging battery:

https://www.kernel.org/doc/Documentation/ABI/testing/sysfs-class-power

Previously, the battery module was parsing them as unsigned longs,
resulting in weird integer overflow behavior from negative rates. This
commit changes the logic to parse rates as potentially signed and handle
that properly.

As I noted in #3219, some changes at some point between the 3.7.2
release and today changed _one_ of the lines that parses the discharge
rate to use signed longs. However, this line actually introduces a new
opportunity for UB:

- It parses the number using strtol, which uses in-band error signaling
  and can produce LONG_MIN on error.
- It does not check the error.
- It passes the result to std::abs, which is _undefined_ when applied to
  LONG_MIN on two's-complement systems. In practice, on normal systems,
  abs(LONG_MIN) == LONG_MIN
- It then bitwise-casts that possibly-still-negative number to
  ULONG_MAX.
- This makes the following math unhappy.

This is a product of two's-complement integers having asymmetrical range
(there are more negative numbers than positive). To avoid this category
of error, I've switched all the relevant code to using floats, and
specifically switched parsing to the C++17 stof function that throws on
error instead of returning a nonsense value. Since the module is already
using exceptions to indicate problems, this seemed like a cleaner
solution.

Along the way, I noticed that a manually sized 16-byte buffer was being
mishandled with very large discharge rate results, of the sort you get
from integer overflows. The specific mishandling would wind up padding
the buffer with NUL characters, out to the correct length of the number,
but only formatting the first 16 characters of it. These NUL characters
wound up getting inserted into the template result and fed to the tag
parser, which interprets NUL as a terminator and rejects the string as
malformed. Thiscommit changes that to a dynamically sized buffer using
std::stringstream, preventing this class of error.

This module is trying to be compatible with battery drivers that present
capacity and rate in units of charge/current (Ah, A) _or_ units of
energy/power (Wh, W). Previously, it made this decision on a
file-by-file basis, so it would accept e.g. a charge_now file for charge
status (in Ah) but an energy_full file for the battery's max (in Wh).
While unlikely, this would result in nonsense results. This commit
alters the logic to choose one or the other and reject such
configurations as incompatible. This change is not strictly related to
the consumption/remaining math fix, but does make reasoning about it a
_lot_ easier. I've added some comments about the units to assist future
analysis of the code.

Fixes #3219